+++ tests/essence_equipment.test.js
@@ -78,6 +78,7 @@ function createHarness() {
     "js/objects/Game_Actor.js",
     "js/objects/Game_Party.js",
     "js/windows/Window_ListViewport.js",
+    "js/windows/Window_ActorNavigator.js",
     "js/windows/Window_Essence.js",
   ]
     .map((relativePath) =>